requirements.txt
# Fraclaw v1.2 Core Requirements

# Core Framework
fastapi
uvicorn
python-socketio
loguru
python-dotenv
python-multipart
aiohttp
httpx

# Messaging Platforms
python-telegram-bot
discord.py

# AI & LLM (Multi-Provider)
openai
anthropic
tiktoken
pydantic

# Memory & Knowledge
chromadb

# Tools - Search & Scraping
duckduckgo-search
beautifulsoup4
readability-lxml
lxml

# Tools - Documents
python-docx
reportlab
openpyxl
python-pptx

# Voice, Audio & Safety
edge-tts
send2trash
faster-whisper
